+++ modules/impex/ui/src/sap/commerce/toolset/impex/options/ImpExProjectSettingsConfigurableProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,12 +22,18 @@ import com.intellij.ide.projectView.ProjectView
 import com.intellij.openapi.options.BoundSearchableConfigurable
 import com.intellij.openapi.options.ConfigurableProvider
 import com.intellij.openapi.project.Project
+import com.intellij.openapi.ui.DialogPanel
+import com.intellij.ui.dsl.builder.AlignX
 import com.intellij.ui.dsl.builder.bindSelected
 import com.intellij.ui.dsl.builder.panel
 import com.intellij.ui.layout.selected
 import sap.commerce.toolset.i18n
+import sap.commerce.toolset.impex.file.ImpExFileType
+import sap.commerce.toolset.impex.ui.components.ImpExQuoteStringExclusionsListPanel
 import sap.commerce.toolset.isHybrisProject
+import sap.commerce.toolset.settings.state.ImpExQuoteStringExclusion
 import sap.commerce.toolset.settings.yDeveloperSettings
+import sap.commerce.toolset.ui.previewEditor
 import javax.swing.JCheckBox
 
 class ImpExProjectSettingsConfigurableProvider(private val project: Project) : ConfigurableProvider() {
@@ -42,82 +48,133 @@ class ImpExProjectSettingsConfigurableProvider(private val project: Project) : C
         private val developerSettings = project.yDeveloperSettings
         private val mutable = developerSettings.impexSettings.mutable()
         private var originalGroupLocalizedFiles = mutable.groupLocalizedFiles
-
+        private val originalQuoteStringExclusions = mutableListOf<ImpExQuoteStringExclusion>()
+        private lateinit var wrapStringExclusionsListPanel: ImpExQuoteStringExclusionsListPanel
         private lateinit var documentationEnableCheckBox: JCheckBox
 
-        override fun createPanel() = panel {
-            row {
-                checkBox("Group localized ImpEx files")
-                    .bindSelected(mutable::groupLocalizedFiles)
-            }
+        override fun createPanel(): DialogPanel {
+            wrapStringExclusionsListPanel = ImpExQuoteStringExclusionsListPanel(project)
 
-            group("Data Edit Mode") {
+            return panel {
                 row {
-                    checkBox("First row is header")
-                        .bindSelected(mutable.editMode::firstRowIsHeader)
+                    checkBox("Group localized ImpEx files")
+                        .bindSelected(mutable::groupLocalizedFiles)
                 }
-                row {
-                    checkBox("Trim whitespace")
-                        .bindSelected(mutable.editMode::trimWhitespace)
-                }
-            }.rowComment("This functionality relies and expects that 'intellij.grid.plugin' is available and enabled.")
 
-            group("Code Completion") {
-                row {
-                    checkBox("Show inline type for reference header parameter")
-                        .comment(
+                group("Data Edit Mode") {
+                    row {
+                        checkBox("First row is header")
+                            .bindSelected(mutable.editMode::firstRowIsHeader)
+                    }
+                    row {
+                        checkBox("Trim whitespace")
+                            .bindSelected(mutable.editMode::trimWhitespace)
+                    }
+                }.rowComment("This functionality relies and expects that 'intellij.grid.plugin' is available and enabled.")
+
+                group("Code Completion") {
+                    row {
+                        checkBox("Show inline type for reference header parameter")
+                            .comment(
+                                """
+                                    When enabled, parameter Type and all its extends will be available as suggestions.<br>
+                                    Sample: <code>principal(<strong>Principal.</strong>uid)</code>
+                                    """.trimIndent()
+                            )
+                            .bindSelected(mutable.completion::showInlineTypes)
+                    }
+                    row {
+                        checkBox("Automatically add '.' char after inline type")
+                            .comment(
+                                """
+                                    When enabled and '.' char is not present, it will be injected automatically
+                                    """.trimIndent()
+                            )
+                            .bindSelected(mutable.completion::addCommaAfterInlineType)
+                    }
+                    row {
+                        checkBox("Automatically add '=' char after type and attribute modifier")
+                            .comment(
+                                """
+                                    When enabled and '=' char is not present, it will be injected automatically.<br>
+                                    In addition to that, code completion will be automatically triggered for modifier values.
+                                    """.trimIndent()
+                            )
+                            .bindSelected(mutable.completion::addEqualsAfterModifier)
+                    }
+                }
+                group("Documentation") {
+                    row {
+                        documentationEnableCheckBox = checkBox("Enable documentation")
+                            .bindSelected(mutable.documentation::enabled)
+                            .component
+                    }
+                    row {
+                        checkBox("Show documentation for type")
+                            .comment(
+                                """
+                                    When enabled short description of the type will be shown on-hover as a tooltip for type in the header or sub-type in the value line.
+                                """.trimIndent()
+                            )
+                            .bindSelected(mutable.documentation::showTypeDocumentation)
+                            .enabledIf(documentationEnableCheckBox.selected)
+                    }
+                    row {
+                        checkBox("Show documentation for modifier")
+                            .comment(
+                                """
+                                    When enabled short description of the modifier will be shown on-hover as a tooltip for type or attribute modifier in the header.
+                                """.trimIndent()
+                            )
+                            .bindSelected(mutable.documentation::showModifierDocumentation)
+                            .enabledIf(documentationEnableCheckBox.selected)
+                    }
+                }
+
+                group("Wrap Value Strings Exclusion Rules", false) {
+                    row {
+                        previewEditor(project, ImpExFileType) {
                             """
-                            When enabled, parameter Type and all its extends will be available as suggestions.<br>
-                            Sample: <code>principal(<strong>Principal.</strong>uid)</code>
+                                UPDATE Title; code[unique = true]; name[lang = en]
+                                # Before inspection rule
+                                ; test ; Not wrapped name
+                                # After inspection rule
+                                ; test ; "Not wrapped name"
                             """.trimIndent()
-                        )
-                        .bindSelected(mutable.completion::showInlineTypes)
-                }
-                row {
-                    checkBox("Automatically add '.' char after inline type")
-                        .comment(
-                            """
-                            When enabled and '.' char is not present, it will be injected automatically
+                        }
+                            .comment(
+                                """
+                                    It is preferable to wrap any values in quotes for non-unique <code>java.lang.String</code> and <code>localized:java.lang.String</code> item attributes.<br>
+                                    This setting extends <code>[y] Quote value strings</code> inspection rule with exclusion of specific item attributes.<br>
+                                    New exclusion rules can be managed here or added from an ImpEx file via quick-fix on a specific attribute.
                             """.trimIndent()
-                        )
-                        .bindSelected(mutable.completion::addCommaAfterInlineType)
-                }
-                row {
-                    checkBox("Automatically add '=' char after type and attribute modifier")
-                        .comment(
-                            """
-                            When enabled and '=' char is not present, it will be injected automatically.<br>
-                            In addition to that, code completion will be automatically triggered for modifier values.
-                            """.trimIndent()
-                        )
-                        .bindSelected(mutable.completion::addEqualsAfterModifier)
-                }
-            }
-            group("Documentation") {
-                row {
-                    documentationEnableCheckBox = checkBox("Enable documentation")
-                        .bindSelected(mutable.documentation::enabled)
-                        .component
-                }
-                row {
-                    checkBox("Show documentation for type")
-                        .comment(
-                            """
-                            When enabled short description of the type will be shown on-hover as a tooltip for type in the header or sub-type in the value line.
-                        """.trimIndent()
-                        )
-                        .bindSelected(mutable.documentation::showTypeDocumentation)
-                        .enabledIf(documentationEnableCheckBox.selected)
-                }
-                row {
-                    checkBox("Show documentation for modifier")
-                        .comment(
-                            """
-                            When enabled short description of the modifier will be shown on-hover as a tooltip for type or attribute modifier in the header.
-                        """.trimIndent()
-                        )
-                        .bindSelected(mutable.documentation::showModifierDocumentation)
-                        .enabledIf(documentationEnableCheckBox.selected)
+                            )
+                            .align(AlignX.FILL)
+                    }
+
+                    row {
+                        cell(wrapStringExclusionsListPanel)
+                            .onIsModified { originalQuoteStringExclusions != wrapStringExclusionsListPanel.data }
+                            .onApply {
+                                mutable.quoteStringExclusions.apply {
+                                    clear()
+                                    addAll(wrapStringExclusionsListPanel.data)
+                                }
+                                originalQuoteStringExclusions.apply {
+                                    clear()
+                                    addAll(wrapStringExclusionsListPanel.data)
+                                }
+                            }
+                            .onReset {
+                                originalQuoteStringExclusions.apply {
+                                    clear()
+                                    mutable.quoteStringExclusions.forEach { add(it.copy()) }
+                                }
+                                wrapStringExclusionsListPanel.data = originalQuoteStringExclusions
+                                    .map { it.copy() }
+                            }
+                            .align(AlignX.FILL)
+                    }
                 }
             }
         }