+++ modules/shared/ui/src/sap/commerce/toolset/ui/Ui.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -19,25 +19,21 @@
 package sap.commerce.toolset.ui
 
 import com.intellij.openapi.Disposable
-import com.intellij.openapi.observable.util.addItemListener
+import com.intellij.openapi.application.invokeLater
+import com.intellij.openapi.fileEditor.FileEditor
+import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.observable.util.addKeyListener
-import com.intellij.openapi.ui.ComboBox
+import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.util.Disposer
-import com.intellij.ui.components.JBScrollPane
-import com.intellij.ui.dsl.builder.Align
+import com.intellij.ui.ClientProperty
+import com.intellij.ui.EditorNotificationPanel
 import com.intellij.ui.dsl.builder.Cell
-import com.intellij.ui.dsl.builder.panel
 import com.intellij.util.asSafely
-import com.intellij.util.ui.JBEmptyBorder
-import java.awt.Color
-import java.awt.Font
-import java.awt.event.ItemListener
+import com.intellij.util.ui.JBUI
 import java.awt.event.KeyListener
 import java.awt.event.MouseListener
 import javax.swing.JComponent
 import javax.swing.JTree
-import javax.swing.ScrollPaneConstants
-import javax.swing.border.Border
 import javax.swing.event.TreeModelListener
 import javax.swing.event.TreeSelectionListener
 import javax.swing.tree.DefaultMutableTreeNode
@@ -45,32 +41,25 @@ import javax.swing.tree.TreePath
 import kotlin.reflect.KClass
 import kotlin.reflect.safeCast
 
-fun scrollPanel(content: JComponent, horizontalScrollBarPolicy: Int = ScrollPaneConstants.HORIZONTAL_SCROLLBAR_AS_NEEDED) = panel {
-    row {
-        scrollCell(content)
-            .align(Align.FILL)
-            .resizableColumn()
-            .applyToComponent {
-                (this.parent.parent as? JBScrollPane)?.apply {
-                    this.horizontalScrollBarPolicy = horizontalScrollBarPolicy
-                    border = JBEmptyBorder(0)
-                }
-            }
+fun <T> DialogWrapper.ifOk(onOk: () -> T): T? = if (this.showAndGet()) onOk()
+else null
 
-    }.resizableRow()
+fun DialogWrapper.repackDialog() {
+    invokeLater {
+        peer.window?.pack()
+    }
 }
 
-fun <T : Any> TreePath.pathData(clazz: KClass<T>): T? = lastPathComponent
-    .asSafely<DefaultMutableTreeNode>()
-    ?.userObject
-    ?.let { clazz.safeCast(it) }
-
-fun <J : JComponent> Cell<J>.border(border: Border?): Cell<J> = this.apply { component.border = border }
-fun <J : JComponent> Cell<J>.background(background: Color?): Cell<J> = this.apply { component.background = background }
-fun <J : JComponent> Cell<J>.opaque(opaque: Boolean): Cell<J> = this.apply { component.isOpaque = opaque }
-fun <J : JComponent> Cell<J>.font(font: Font): Cell<J> = this.apply { component.font = font }
-fun <J : Any> Cell<ComboBox<J>>.addItemListener(parentDisposable: Disposable? = null, listener: ItemListener): Cell<ComboBox<J>> = this
-    .apply { component.addItemListener(parentDisposable, listener) }
+fun DialogWrapper.banner(
+    text: String,
+    status: EditorNotificationPanel.Status = EditorNotificationPanel.Status.Info
+) = EditorNotificationPanel(null as FileEditor?, status).apply {
+    this.text = if (text.contains("<html>")) text else "<html>$text</html>"
+    border = JBUI.Borders.compound(
+        ClientProperty.get(this, FileEditorManager.SEPARATOR_BORDER),
+        border
+    )
+}
 
 fun <T : KeyListener, J : JComponent> Cell<J>.addKeyListener(parentDisposable: Disposable? = null, listener: T): Cell<J> = this
     .apply { component.addKeyListener(parentDisposable, listener) }
@@ -96,3 +85,8 @@ fun <T : MouseListener> JTree.addMouseListener(parentDisposable: Disposable? = n
 private fun Disposable.whenDisposed(onDispose: () -> Unit) {
     Disposer.register(this) { onDispose() }
 }
+
+fun <T : Any> TreePath.pathData(clazz: KClass<T>): T? = lastPathComponent
+    .asSafely<DefaultMutableTreeNode>()
+    ?.userObject
+    ?.let { clazz.safeCast(it) }